refactor: change directory and file names to plural
