+++ functions/src/listeners/handleOverflowAction.ts
@@ -4,7 +4,7 @@ import {
   SlackActionMiddlewareArgs,
 } from "@slack/bolt";
 import * as functions from "firebase-functions";
-import { RotationMessage, RotationModal } from "../component";
+import { RotationMessage, RotationModal } from "../components";
 import { getUserNameDict } from "../services/getUserNameDict";
 
 export const handleOverflowAction: Middleware<
