+++ functions/src/models/__tests__/rotation.test.ts
