+++ functions/src/models/rotation.ts
