+++ functions/src/models/schedule.ts
